+++ src/core/dbpf-entry.ts
@@ -6,7 +6,8 @@ import type { Class, Constructor } from 'type-fest';
 import type { InspectOptions } from 'node:util';
 import WriteBuffer from './write-buffer.js';
 import Stream from './stream.js';
-import { getConstructorByType, getTypeLabel, hasConstructorByType } from './helpers.js';
+import { getTypeLabel } from './helpers.js';
+import { getConstructorByType, hasConstructorByType } from './file-classes-helpers.js';
 import type DBPF from './dbpf.js';
 import FileClasses from './file-classes.js';
 import FileType from './file-types.js';
